Create random_study.py
You may notice that the output list may contain duplicated. This happens because of the priority score that you have given the subjects. Python is determining the duplicates have a high enough score to appear more than once. In such cases, if you have asked for 4 subjects for 4 study sessions, and 1 subject appears twice, it means you should spend 2 sessions studying that subject and 1 each for the other two.